requirements.txt
Django==2.2
django-appconf==1.0.3
django-cors-headers==2.5.2
django-mediumeditor==1.0.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
